move configuration inside a folder
