+++ components/configuration/NetPyNESimConfig.js
@@ -7,9 +7,9 @@ import Tooltip from 'material-ui/internal/Tooltip';
 import FlatButton from 'material-ui/FlatButton';
 import Card, { CardHeader, CardText } from 'material-ui/Card';
 
-var PythonControlledCapability = require('../../js/communication/geppettoJupyter/PythonControlledCapability');
+var PythonControlledCapability = require('../../../../js/communication/geppettoJupyter/PythonControlledCapability');
 var PythonControlledTextField = PythonControlledCapability.createPythonControlledComponent(TextField);
-var PythonControlledSelectField = PythonControlledCapability.createPythonControlledComponent(SelectField);
+
 
 const styles = {
   populationCard: {
